feat(ui): refined Slack-style message rows + extract AttachmentDisplay (#565)
